# docs/specs/S01-openapi.yaml
openapi: 3.1.0
info:
  title: Baby Growth Record API
  version: 2.3.0
  description: API specification for the baby growth record SaaS platform. IDs are decimal strings at API boundary and stored as bigint in database; time fields use UTC millisecond timestamps.
servers:
  - url: /api/v1
...
paths:
  /auth/wechat/login:
    post:
      tags: [auth]
      summary: Login with WeChat (phone-based identity)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/password/login:
    post:
      tags: [auth]
      summary: Login with phone and password (auto-register on first login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke current access/refresh token
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
...
  /ai/query:
    post:
      tags: [ai]
      summary: Query AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiQueryRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiQueryResponse'
  /ai/summary:
    post:
      tags: [ai]
      summary: Generate AI summary (daily/weekly/monthly)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiSummaryRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
...
components:
...
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
...
    FamilyRole:
      type: string
      enum: [owner, caregiver, viewer]
...
    WechatLoginRequest:
      type: object
      required: [code, phone]
      properties:
        code:
          type: string
        phone:
          type: string
          pattern: '^1\d{10}$'
        encrypted_phone_data:
          type: string
        iv:
          type: string
    PasswordLoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          pattern: '^1\d{10}$'
        password:
          type: string
          minLength: 8
          maxLength: 128
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9]+$'
        display_name:
          type: string
    Family:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9]+$'
        name:
          type: string
        role:
          $ref: '#/components/schemas/FamilyRole'
        timezone:
          type: string
...
    WechatLoginData:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        families:
          type: array
          items:
            $ref: '#/components/schemas/Family'
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/WechatLoginData'
    RefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
...
    AiSummaryRequest:
      type: object
      required: [family_id, baby_id, summary_type]
      properties:
        family_id:
          type: string
          pattern: '^[0-9]+$'
        baby_id:
          type: string
          pattern: '^[0-9]+$'
        summary_type:
          type: string
          enum: [daily, weekly, monthly]
        summary_date:
          type: integer
          format: int64
...
    AiQueryRequest:
      type: object
      required: [family_id, baby_id, question]
      properties:
        family_id:
          type: string
          pattern: '^[0-9]+$'
        baby_id:
          type: string
          pattern: '^[0-9]+$'
        question:
          type: string
    AiQueryResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                answer:
                  type: string
                window_start:
                  type: integer
                  format: int64
                window_end:
                  type: integer
                  format: int64
                disclaimer:
                  type: string